refactor: merge brushinfo and brushctl
